+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/content/abstracts/MediaGroupContent.kt
@@ -1,8 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia.MediaGroupMemberInputMedia
-import com.github.insanusmokrassar.TelegramBotAPI.types.files.abstracts.TelegramMediaFile
 
-interface MediaGroupContent<T : TelegramMediaFile> : MediaContent<T>, MessageContent {
+interface MediaGroupContent : MediaContent, MessageContent {
     fun toMediaGroupMemberInputMedia(): MediaGroupMemberInputMedia
 }